Introduced basic server-side form validation for required fields
Ideally, we'll duplicate all error validation on the server, so we're not reliant on the client-side validation. E.g., we might validate regex, minlength, maxlength, &c. That said, _at minimum_, we need to validate that required fields are set. This update accomplish that. This, too, will require some care from @ktrunkey to fix formatting.