Use derived `Key` if appropriate
If the topic derives from another topic, and the `Key` is `null`, then default to the `Key` of the derived topic. In the legacy setup, this would autogenerate a new `Key`—which was almost never the desired approach. This is much more intuitive in _most_ use cases. (One exception being if it's deriving from a topic in the same parent; an issue we'll address shortly.)