Resolved issue with `IsNew` updating the parent topic
Perhaps confusingly, `CurrentTopic` is set to the _parent_ topic when `IsNew`. That's because the user is currently in the parent topic, and hasn't yet created the new topic. Perhaps we should revisit the naming conventions to avoid ambiguity in this case (which, it should be noted, only occurs in the editor). Regardless, this mandates updating the code to modify the local `topic` object instead—which _may_ be set to `CurrentTopic`, but may _instead_ be set to a new `Topic` if `IsNew`.