Introduced functionality to cross-validate all tabs
By default, **jQuery Validate** only validates the fields in the _current_ tab. That's because the fields in the remaining tabs are hidden, and hidden fields don't get validated. This is fixed by introducing `ignore` into the `validate()` call. In addition, an `invalidHandler` is introduced in order to ensure that tabs with errors are properly marked. Notably, an asterisk is appended to each one, and a new class, `error`, is added to the link itself. Shout out to Gabriel Buzzi Venturi for the useful example on **CodePen**. As with other validation-related features, this will need some care from @ktrunkey.