Resolved issue with potentially empty `TopicID`
If the `TopicID` attribute is found (expected) _but_ is empty (usually the case) then `Int32.Parse()` would fail. Mitigated that by checking `IsNullOrEmpty()` prior to calling `Int32.Parse()`. This is a touch clumsy with three distinct variables, but it's simpler than having a number of nested conditions written out with `if/then` statements.