Missed replaces for `absolutePath`
It makes sense to reuse the `absolutePath` variable wherever possible. But this is especially important here since the implementation of `absolutePath` in #3c79ffe relies on a different source property for calculating the base application path. This change was actually complete previously, but somehow didn't get picked up by git.