Introduced a `StandardEditorComposer`
I'm not in love with this approach, and may reevaluate it prior to launch, but the `StandardEditorComposer` provides a basic implementation for aiding in the composition of dependency graphs around the standard out-of-the-box view components required by the editor, and thus helps eliminate the need for caller applications to know and register every single attribute type view component. That said, this approach introduces some tradeoffs; namely, it presumes certain lifestyles, and mandates an `ITopicRoutingService` be created, even though only one of the view components (`FilePathViewComponent`) even requires it. That said, this assumptions should be valid in the vast majority of use cases.