Moved infrastructure components to `Internal`
Since these libraries are not needed outside of the editor, and we don't anticipate the need for implementers to derive them or use them as part of the project composition, we mark them as `internal`. We can always (re)expand access later, should a need be discovered. In the meanwhile, this helps minimize our public API. (In practice, we might be able to reduce access to many of the models, as well, but that will require establishing a trust between the assemblies, since they're stored in separate projects.)