Ensure relationship values are preset
When values are modified, **ExtJS** updates a hidden backing field with a comma-delimited list of `Topic.Id`s. It doesn't preset these values on load, however. As such, if a topic is loaded, no relationships are modified, and then the topic is saved, it will remove all relationships, since there aren't any `Topic.Id`'s in the backing field. Oops. To prevent this, we're ensuring that the default value is being set for `ModelType.Relationhip`—even though it's _also_ set (at least in the user interface) via the call to the `/JSON` action.