Honor preexisting `Key` value for `IsNew`, `DerivedTopic`
If the user doesn't set the `Key`, but instead sets the topic to derive from another topic, the existing topic creation logic will recognize that and set the `Key` to the `DerivedTopic.Key`, which is intuitive from a user perspective, and avoids complicated form validation logic for ensuring the `Key` is set. That said, this logic was being overwritten by the default logic for setting attribute values, as it would come across the `Key` value (which would be empty in that scenario) and overwrite the inherited value. This can be mitigated by ignoring the attribute logic for `Key` in the `IsNew` scenario. (It is retained for update scenarios.)