Only show `InheritedValue` for derived topics.
Previously, the `InheritedAttributes` (which populate the `InheritedValue` property on the view model) were being set to inherit from parent attributes. That's misleading, since we very rarely inherit attribute values from the parent; it implies a default that likely doesn't exist in most scenarios. Instead, the logic is modified to grab the value from the current topic only. Since, in practice, this value will be overwritten with an _actual_ `value` attribute in the interface, that means this will only show values inherited from a `DerivedTopic`. This logic is disabled for `IsNew` as, otherwise, it would default to the values of the parent topic for `IsNew` scenarios, since the new topic isn't yet available. (This could be made a bit more explicit by caling `GetValue()` off of the `DerivedTopic`, but that complicates the logic and has no impact on the front end.)