Made the value of radio buttons nullable
By allowing `null` values for radio buttons, we allow them to inherit values from derived (or, potentially, parent) topics, and also allow implicit defaults to be implemented in the code. As part of this, replaced the `IsSelected()` helper method on the `BooleanAttributeViewModel` with separate `IsTrue()` and `IsFalse()` methods. This simplifies the logic for handling nulls in the view itself.