Uncommented calls to `ITopicRepository` in preparation for deployment
During testing, destructive calls to `ITopicRepository` were commented out. This allowed us to test the editor code and interface against a real data set without worrying about making accidental changes, or needing to constantly restore back to a stable condition. Now that we're preparing to release this version, we are able to (finally) uncomment these, thus allowing the modifications drafted in the editor to make changes to the target database. These correspond with integration tests of `Save()`, `Move()`, `Rollback()`, and `Delete()` prior to pushing this commit.