Corrected bug with removing existing items
Because existing values are derived from the `value` attribute of the hidden HTML backing field, they are interpreted as strings. The `id` field is an int. As a result, when a `push` or `remove` is performed, it doesn't see the existing value and the new value as the same. This can result in old values not getting removed—or, worse, getting duplicated (in the case that a relationship is toggled), and causing errors down stream. This is easily fixed by treating the backing field as a string array, and converting the `id` to a string before `push`ing or `remove`ing it.