Bumped *Topics Library* to latest patch
This version accounts for removing unnecessary variables on the `CreateTopic` and `UpdateTopic` stored procedure calls. The database schema is also updated to ensure that `ExtendedAttributes` are only created if they differ from the previous version (often they may not, if e.g. changing indexed metadata such as `IsHidden` or the `Key`).