Refactor delete button markup and confirmation
The **jQuery** script now operates off an inline `data-title` and `href`, instead of dynamically creating the script in `Edit.cshtml`. This allows the confirmation code to be moved to `UserInterface.js`, for consistency with e.g. the modal confirmation code. Further, modified `disableDelete` to be nullable, so the attribute will be removed by **Razor** if it's not set. Finally, ensures `isModal` is being relayed as part of the `href`. Since this required changing the `<button />` to an `<a />`, I also modified the **SCSS** selector for buttons to be more generic (i.e., to select both `button` and `.btn`).