Refactored `Ignia.Topics` to `OnTopic`
While **OnTopic** continues to be maintained by **Ignia**, it has enough libraries that it makes sense to separate out into a new organization, **OnTopic**. In preparation for that, we're simplifying the namespaces (and project names) down from `Ignia.Topics.*` to simply `OnTopic.*`. This affects nearly every file in the project, for obvious reasons. As part of this, we're also upgrading to the final **OnTopic Library 4.0.0** release, which has already applied these changes.