Introduced compatibility layer for `DefaultConfiguration`
In the final release of **OnTopic 4.0.0**, `DefaultConfiguration` and `Configuration` were removed from the core `AttributeDescriptor` class, and moved to the **OnTopic-WebForms**' version of `ConfigurableAttributeDescriptor`, since the `DefaultConfiguration` is only needed for backward compatibility with the **OnTopic-Editor-WebForms** client. That said, to aid in implementors transitioning from **OnTopic-Editor-WebForms** to **OnTopic-Editor-AspNetCore**, we _want_ to maintain a _basic_ level of temporary backwards compatibility. That said, while this provides a convenience for migrations, we don't want this to be confused as a permanent part of the API, so we're marking the impacted members as `[Obsolete]`—and temporarily supressing the warnings within the `Mvc` project since this impact is well-understood.