Removed interstitial `/Areas/Editor` folders
The original **Razor Class Library** template established an `/Areas` folder. That's probably a good practice in general. In our case, however, we don't want (or expect to need) the editor to be in its own area—the default `/OnTopic` route should be sufficiently unique (though, admittedly, `EditorController` could be a _potential_ issue). Currently, having these folders buried two levels doesn't _hurt_ anything, but it _implies_ the use of an area ("Editor") that's not there, and it also makes traversing the tree an extra pain due to the unnecessary depth.