Fixed path to `/_content` area
With the rename(s) of `Ignia.Topics.Editor.Mvc` to (ultimately) `OnTopic.Editor.AspNetCore`, the JavaScript file reference location is broken. Typically, we don't have (non-relative) file paths in JavaScript, but **CKEditor** is an exception, since it dynamically loads files. (Also fixed an errant tab while I was at it.)