Ensure background doesn't scroll with page
Previously, while the `#Toolbar` remained fixed, the background of the `#FormArea` would scroll underneath the toolbar, giving it a white background. This gives the impression that the background is scrolling, while the contents are not. The root cause is that the `#Toolbar` itself is just inheriting its background color from the underlying panel—which is actually the `body` to begin with. To resolve this, the body's `background-color`—`$grey-dark` or `#333`—is being applied directly to `#Toolbar`.