Allow layout to be hidden on the `TopicListViewComponent`
The `TopicListViewComponent` is unique in that it's not used _exclusively_ as an attribute field, as other view components in the editor are. Instead, it is also used in the navigation and within the `NestedTopicListViewComponent` in order to allow new topics to be created. In the future, we may want to separate this out given how divergent the functionality has become. In the meanwhile, however, we at least want to allow the `_Layout.cshtml` (which displays the label and injects hidden fields) to be conditionally removed. We can infer this state based on whether or not the `target-url` (`targetUrl`) parameter is passed; if it is, it is being embedded for navigation; otherwise it is not. Again, we'll probably want to revisit how we handle this in the future, but for now we want the ability to hide that layout. This also necessitates increasing the `margin` below the `TreeView`.