Enabled `UseFullSemVerForNuGet` for consistency in versioning
The `UseFullSemVerForNuGet` flag is enabled on the main **OnTopic Library 4.0.0** project, which prevents the need for e.g. padding the build or release numbers. By enabling it here, we allow more consistent version conventions between products. (There are other benefits to enabling this flag, but that's the main driving force for enabling it right now.)