Key view component off of `EditorType`
Technically, this should always end up being the same value as keying off `ContentType`, _but_ it's more intuitive and consistent to use `EditorType`, since that's explicitly what it represents. Further, it also provides more flexibility if, in the future, we want to differentiate the attribute name from the content type, for some reason.