Help _Topic Information_ fill available space
I don't like this approach, and would rather e.g. `@extend .col-md-3`, for example. Since we don't (yet) have **Bootstrap** integrated with our **SASS** structure, however, that's not possible. And it may not matter even if it were, since this is a `fixed` column, which brings it outside the context of its parent. Given that, extending this to 15% brings it close(r) to the mark, while constraining it from excess via the `max-width`. That said, this remains an approximation that is imprecise, and really would benefit from a new approach.