Update Edit.cshtml
Marked the `callouts` to hide if the screen size is smaller than a `lg`, which should help e.g. the tablet display. (We don't expect people to use the **OnTopic Editor** on a mobile phone, but tablets should be considered.) In that scenario, we also make the main form wider, to adopt the space left behind.