Removed gutters for `IsModal`
Out of the box, the **Bootstrap** `.row` has margins of `0px -15px` to compensate for the padding of the gutters, and allow the row itself to fill the space. That said, when our content area is occupying the entirety of the body, as is the case for the modal window, this causes the `.row` to be 30px wider than the window—regardless of what size we set the window to. By applying `no-gutters` in this situation, we instruct **Bootstrap** that the negative margins aren't desired here, and to remove them. This also allows us to get rid of the previous extended width applied to the `.modal-lg` state.