Resolve bug which could result in malformed modal URLs
The previous `targetUrl` definition _assumed_ that the current URL ended in a `/` and, thus, didn't add one as a delimiter between the current `Action()` and the `AttributeDescriptor.Key`. If the current URL does _not_ end in a `/`, this would result in the modal window loading with a malformed URL, thus resulting in an error.