Introduced new `ContentTypeListViewComponent`
Previously, the `TopicListViewComponent` was awkwardly trying to satisfy two scenarios: 1) An editor attribute which dynamically looks up a list of topics, and 2) A navigation control that displays a list of content types. This required a lot of conditional code to account for different scenarios. To help mitigate that, the latter scenario has been extracted from the `TopicListViewComponent` (and related files) and moved into a new `ContentTypeListViewComponent`. I don't love this implementation—mostly because there are still optional parameters that must be supplied via declarative markup—but it at least helps differentiate the use cases. As part of this, implemented the new `ImplicitlyPermitted` attribute on `ContentTypeDescriptor`, so that it's easier to restrict what content types are otherwise available on any given page.