Remove _Delete_ if `DisableDelete`
This was previously set up to disable the delete button. That's actually preferrable. That only works, however, if the _Delete_ button is an actual `<button />` element. Since this "button" is actually an `<a />` element, the `disable` attribute isn't supported. There are ways to get this to work (likely removing the `href` and any `onclick` handlers) but those are more involved. For now, a quick fix is just to get rid of the button entirely. (Which is actually how we handle the _Versions_, _Cancel_, and even the _Delete_ button in the `IsNew` state, so perhaps it's actually preferrable?)