Allow permitted content types to be defined on `Category` topics
Traditionally, the types of topics that can be created underneath another topic are defined on a topic's content type, via the _Permitted Content Types_ (`ContentTypes`) attribute. For containers, however, it's useful to define this on the topic itself. This way, when we setup a container for organizational purposes, we can specify what type of content types are to be created under that _instance_—as opposed to creating an e.g. `FormContainer` to limit child topics to the `Form` content type. This obviously doesn't provide any real security, but the goal is to fine-tune the user experience so editors don't need to pick through a list for the one type expected, nor do fairly specific types (such as `Form`) need to be set to _implicitly permitted_, thus displaying them well beyond their natural boundaries.