Move views back into `/Areas`
This effectively reverts a portion of #f618fc8, which removed the interstitial `/Areas/Editor` folders in order to flatten the folder structure. It turns out that views in a **Razor Class Library** are not namespaced, unlike `/wwwroot`, and are thus subject to being subsequently overwritten by host sites embedding the library. This is especially concerning for `/Views/Shared` where common file names, such as `_Layout.cshtml`, could easily take priority. This is actually a useful behavior in some contexts, as it allows more flexibility on overriding elements. But it's peculiar that Microsoft applied this to `/wwwroot` but not `/Views` (e.g., by creating some type of assembly based namespacing convention). Having the files buried a couple additional levels is less than ideal organizationally, but at least we can keep the `.cs` files in (children of) the root directory.