Make `GetEditorViewModel()` generic
Previously, `GetEditorViewModel()` would only return an `EditorViewModel` object. As we introduce additional functionality, however, we're going to want to subclass `EditorViewModel`. By making `GetEditorViewModel()` generic, that allows us to reuse this same method to deliver an `EditorViewModel` _or_ derived type, as needed. That provides a lot more flexibility. Since creating generic types requires a default constructor, however, that also means that we need to remove the constructor from `EditorViewModel` and make the properties writable. (This is actually how we handle _most_ view models, due to how the `ITopicMappingService` operates, so it's also more consistent.)