Add *Font Awesome* icon to _Add Topic_ button
Added a `+` icon to the `ContentTypeList` view component, which is displayed on the left-hand rail and as part of nested topics for creating a new topic. While not strictly necessary, this adds consistency with the implementation of the new _Import_ and _Export_ buttons, which use an up and down arrow icon respectively.