Update GitVersion
Update to the latest version of `GitVersion` task. We had been holding off on this due to some apparent churn and instability in minor versions. With those worked out, however, we can now take advantage of the new bug fixes and features. As an internal (private) dependency, this only affects our build process, and not any downstream consumers of the **NuGet** package.