Set the `CurrentUser` when doing imports
The `Import()` extension method provided by the Data Transfer Library doesn't have access to the `HttpContext` and, thus, can't set the `LastModifiedBy` to the current user if that option is selected via the editor (see #5361a8e). This can be mitigated by injecting that value into the `ImportOptions` object prior to import. Note that this behavior is by design, but it does require this configuration from callers to ensure that it works correctly.