Mitigated issue with non-nullable checkboxes
The `ImportOptions.DeleteUnmatched*` attributes are all `bool`, but are backed by a nullable `bool?` which defaults to `null` and inherits its value based on the `ImportOptions.Strategy` value. Unfortunately, when using the ASP.NET Core `input` tag helper with checkboxes, a hidden field is generated with the value of `false`. This results in the value of these properties _always_ being set to false, and no longer determing their value from the `ImportOptions.Strategy`. We only want that to happen when they're set to `true`. To mitigate that, we're removing the tag helper, and manually defining the `name` attribute. This introduces a risk in the case that the `enum` values change; to help avoid that potentiality in the future, we used `nameof()` on both the `enum` as well as its fields, so a compilation error will occur if those values are altered.