Hack: Delete unmatched topics
While the `Import()` extension method will successfully remove unmatched topics from the topic graph—assuming e.g. `DeleteUnmatchedChildren` is enabled—it doesn't have access to the `ITopicRepository` in order to delete them. The `ITopicRepository` doesn't have a way to detect internal deletions either, however.

In the future, we should evaluate ways of baking this into `ITopicRepository` e.g. by temporarily moving deleted topics to a `Topic.DeletedChildren` collection. For now, however, we can hack around this by taking a snapshot of all topics in the graph before hand, finding the difference afterwards, and deleting those topics that are no longer present.

Technically, this will result in some topics being deleted which are already deleted (e.g., because their parent was already deleted), but, for better or for worse, the `ITopicRepository` handles that situation gracefully (i.e., without throwing an exception).

This is an ugly hack, and especially because it's not intuitive behavior when using the Data Transfer library. It doesn't require too much code, however, and works reliably, so it's an acceptable workaround until a more sophisticated solution can be accounted for in the `ITopicRepository`.