Update to OnTopic 4.2.0
Among some feature updates, OnTopic 4.2.0 notably includes a fix for a bug in how versions are established (OnTopic-Library@90c83d4). This resolves a bug in OnTopic Editor where a user is unable to rollback to a version of a topic that was created prior to the last application reset. This was due to the `VersionHistory` being populated with a higher-precision version timestamp than used in SQL server; once once the the application was reset would the lower-precision value be retrieved from the database. Now, the version is initialized using the lower-precision value so that these are always synchronized.