Perform redirect on`movenode()`
When the current topic—or any of its ascendents—is moved in the `TreePanel`, the current page's URL will no longer be valid. As a result, if the editor then performs an action on that page, such as saving a change, it will fail since it will attempt to post back to the original URL.

To mitigate this, we now perform a redirect in `movenode()` if the `node` is the `currentTopic` or an ascendent of it. This will put the user on the same topic, but in its new location.