Perform `reload()` on `movenode()`
When a topic is moved in the `TreePanel`, the topic (and any of its descendants') `webPath` and `path` values will have changed. As such, the `TreePanel` needs to be `reload()`ed so that those are refreshed. Otherwise, if they are selected, they will go to a no-longer-valid location.

This is only done if the entire page is not reloaded (see e6a33fa) as, in that case, the `TreePanel` will be reloaded as part of the page load instead.