Introduced new `OnTopic.Navigation` JavaScript class
Instead of defining the navigation tree inline as part of the `_TreeViewScript.html`, I've moved it to a new `OnTopic.Navigation` class using **ExtJS**'s `Ext.extend()` method. This doesn't change the functionality, but it allows the JavaScript to be minimized, loaded, and cached as part of the main `Scripts.js`. It also makes it easier to format the script for readability by breaking up the event handlers into object methods.