Remove unused variables
The `oldPosition`, `oldNextSibling`, and `rootTopicId` were used by previous versions of the OnTopic Editor to handle move events. They aren't necessary based on the current implementation, and are vestigal variables. (`rootTopicId` is used to construct the `root` `AsyncTreeNode`; while that node is still necessary, it doesn't actually need an `id` or `text` value since it's no longer displayed.)