Rename `dragTopic` to `validateSourceTopic`
With the `oldPosition` and `oldNextSibling` variables removed (1c9c26d), the `dragTopic` handler is really just being used to validate that the source topic can be drug. Given this, we're able to give it a more descriptive and intuitive name. This way, when it shows up among other event handlers in the events tree, it's clear what its purpose is.