Introduced new `NestedTopics` JavaScript class
The `NestedTopics` control extends the `DraggableTreeView` (09675aa) in order to expose a tree view optimized for rendering a flat list of nested topics. Since nested topics aren't actually hierarchical, the main benefit of reusing the underlying `DraggableTreeView` is the ability to share both rendering and topic moving logic.

The main addition of the `NestedTopic` control is the `nodedragover` event handler, `validateTargetTopic`, which ensures that nested topics aren't drug underneath other topics; they can be _reordered_ but their _parent_ can't be modified.