Updated to comply with Ignia's latest style guide, formatting standards
As part of the lead-up to OnTopic 4.0.0, most of the scripts were updated to use Ignia's latest style guide. For some reason, however, the `TokenizedTopicList` was overlooked, and continued to use legacy comment blocks, and comment block indentation. The file has been updated to use the latest standards, for consistency with other script files.