Introduced new `OnTopic.SelectableTreeView` JavaScript class
Centralized the configuration for the relationships into a new `OnTopic.SelectableTreeView` JavaScript control, similar to what was previously done with `OnTopic.Navigation` (e6a93b7) and `OnTopic.NestedTopics` (33b4d1f).

This makes the view cleaner, keeps the JavaScript in a JavaScript class which is easier to validate and debug, and (as a minor improvement) ensures the code isn't duplicated in the markup if multiple relationships are on a page (rare).