Introduce tolerance for missing content type descriptors
The `ContentTypeListViewComponent` is intended to list content type descriptors allowed to be created under the current content type descriptor. This won't be possible if the content type descriptor can't be found.

This is a rare scenario that should only occur when either a) working with an as-of-yet unsaved in-memory topic graph containing new content types, or b) working with an empty data repository, in which case a default `Root` topic will be created, even though the `Container` content type it points to doesn't yet exist.

While neither of these scenarios are common, and should never happen for established topic databases, we can easily bypass this possibility by returning an empty view model in these cases. This will simply result in the `ContentTypeListViewComponent` being hidden from the interface. In turn, this opens up opportunities for e.g. importing reference schemas via JSON for otherwise empty data repositories.