Ensure a `ContentTypeDescriptor` is always available
The `GetContentType()` method is responsible for looking up the `ContentTypeDescriptor` of the `CurrentTopic` and returning it so that it can be factored into the corresponding `EditorViewModel`. This is then used to generate the editor form.

When working with an empty database, however, this isn't possible. While a `Root` topic with a `Container` content type will be automatically generated, the corresponding `Container` content type descriptor won't be available. Obviously, that means that the `Root` topic can't be edited. But the editor should still render in that case, thus allowing a reference topic schema to be imported via the **OnTopic Data Transfer** library.

By allowing `GetContentType()` to fallback to a default, empty `ContentTypeDescriptor` if the current content type can't be located, it allows this scenario—while still preventing invalid content types from being created, since it won't be able to generate a valid form for the content type without any `ContentTypeDescriptors.AttributeDescriptors`.