Upgraded to preview of OnTopic Library 4.3.0 (build 60)
This resolves an issue with `TopicRepositoryBase.GetContentTypeDescriptors()` throwing an exception the first time it attempts to load data from an empty database. It has been updated with a new `TopicNotFoundException` that is swallowed by `GetContentTypeDescriptors()` in order to accommodate an empty database.

This should also resolve a bug in how the `GetTopicId()` function was handled in the `SqlTopicRepository` which threw an off-target exception when attempting to parse the return code when no topic could be found (again, as would be the case when the database is empty).