Default to `Root`
Previously, we defaulted to `Root/Web`. There are benefits to this, since there shouldn't be a legitimate reason for editors to modify (or even need to be aware of) the `Root`, and most content they _are_ interested in will be under `Root/Web`, so it's handy to have that section of the tree expanded by default. Unfortunately, this conflicts with the ability to handle an empty database.

For now, we'll change the default to `Root`. We may want to determine if we can implement a more intelligent fallback, however, so that it still defaults to `Root/Web`, but can fall back to `Root` if the `Web` container isn't present.