Updated to OnTopic Library preview
This preview release of the **OnTopic Library** includes new features the `Topic.IsNew` property. It also includes considerable improvements to the performance of recursive `Save()`s, which are of particular value to the **OnTopic Data Transfer** library. Notably, `Save()` is now able to detect if any attributes or relationships have changed—and, if they haven't, skip the remainder of the `Save()` entirely.

Obviously, this will need to be updated to the release version once it's available—and prior to this update to the **OnTopic Editor** being released.