Ensure labels aren't treated as values
The `TopicListViewComponent` includes support for a `DefaultLabel`. Previously, the `SelectListItem.Value` was set to `null`. This results in .NET (or, perhaps, HTML) using the label as the value, if no other value is selected. E.g., if the `TopicListAttribute`'s `DefaultValue` is set to "Select a topic…" and no topic is selected, then the value would be set to "Select a topic…". That's not what we want. By using an empty string instead, we mitigate this issue.