Support `RelativeTopicBase` with new topics
Previously, the `TopicList`'s `RelativeTopicBase` would work incorrectly with new topics since `ITopicRepository.Load()` would load the _parent_ context's `UniqueKey`, and thus it would return content relative to the _parent_, not the new topic.

This is mitigated by temporarily creating a placeholder topic in the topic graph that can be used as the relative base. This placeholder topic must be removed from the parent's `Children` collection, as otherwise it will persist in memory, interfere with future attempts to create new children, and potentially even end up getting persisted to the database under certain circumstances (such as a recursive `Save()` triggered by an `Import()`).